pyproject.toml
[tool.poetry]
name = "dash-nuitka-boilerplate"
version = "0.1.0"
description = ""
authors = ["Lukas Tenbrink <lukas.tenbrink@gmail.com>"]
readme = "README.md"
packages = [
    { include = "dnb", from = "src" },
]

[tool.poetry.dependencies]
python = "~3.9"
dash = "^2.17.1"
pandas = "^2.2.2"
nuitka = "^2.3.9"
pywebview = "^5.1"
setproctitle = "^1.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
